fix(api): move `children` endpoint
