+++ bin/inx-chronicle/src/api/stardust/core/responses.rs
@@ -17,16 +17,3 @@ pub struct InfoResponse {
 }
 
 impl_success_response!(InfoResponse);
-
-/// Response of GET /api/core/v2/blocks/{block_id}/children.
-/// Returns all children of a specific block.
-#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct BlockChildrenResponse {
-    pub block_id: String,
-    pub max_results: usize,
-    pub count: usize,
-    pub children: Vec<String>,
-}
-
-impl_success_response!(BlockChildrenResponse);
