+++ bin/inx-chronicle/src/api/stardust/explorer/responses.rs
@@ -79,3 +79,16 @@ pub struct BalanceResponse {
 }
 
 impl_success_response!(BalanceResponse);
+
+/// Response of GET /api/explorer/v2/blocks/{block_id}/children.
+/// Returns all children of a specific block.
+#[derive(Clone, Debug, Eq, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct BlockChildrenResponse {
+    pub block_id: String,
+    pub max_results: usize,
+    pub count: usize,
+    pub children: Vec<String>,
+}
+
+impl_success_response!(BlockChildrenResponse);
