+++ bin/inx-chronicle/src/api/stardust/explorer/routes.rs
@@ -6,7 +6,7 @@ use std::str::FromStr;
 use axum::{extract::Path, routing::get, Extension, Router};
 use chronicle::{
     db::MongoDb,
-    types::stardust::block::{Address, MilestoneId},
+    types::stardust::block::{Address, BlockId, MilestoneId},
 };
 use futures::{StreamExt, TryStreamExt};
 
@@ -15,17 +15,22 @@ use super::{
         LedgerUpdatesByAddressCursor, LedgerUpdatesByAddressPagination, LedgerUpdatesByMilestoneCursor,
         LedgerUpdatesByMilestonePagination,
     },
-    responses::{BalanceResponse, LedgerUpdatesByAddressResponse, LedgerUpdatesByMilestoneResponse},
+    responses::{
+        BalanceResponse, BlockChildrenResponse, LedgerUpdatesByAddressResponse, LedgerUpdatesByMilestoneResponse,
+    },
 };
-use crate::api::{ApiError, ApiResult};
+use crate::api::{extractors::Pagination, ApiError, ApiResult};
 
 pub fn routes() -> Router {
-    Router::new().route("/balance/:address", get(balance)).nest(
-        "/ledger/updates",
-        Router::new()
-            .route("/by-address/:address", get(ledger_updates_by_address))
-            .route("/by-milestone/:milestone_id", get(ledger_updates_by_milestone)),
-    )
+    Router::new()
+        .route("/balance/:address", get(balance))
+        .route("/blocks/:block_id/children", get(block_children))
+        .nest(
+            "/ledger/updates",
+            Router::new()
+                .route("/by-address/:address", get(ledger_updates_by_address))
+                .route("/by-milestone/:milestone_id", get(ledger_updates_by_milestone)),
+        )
 }
 
 async fn ledger_updates_by_address(
@@ -125,5 +130,29 @@ async fn balance(database: Extension<MongoDb>, Path(address): Path<String>) -> A
         total_balance: res.total_balance,
         sig_locked_balance: res.sig_locked_balance,
         ledger_index: res.ledger_index,
+    })
+}
+
+async fn block_children(
+    database: Extension<MongoDb>,
+    Path(block_id): Path<String>,
+    Pagination { page_size, page }: Pagination,
+) -> ApiResult<BlockChildrenResponse> {
+    let block_id = BlockId::from_str(&block_id).map_err(ApiError::bad_parse)?;
+    let mut block_children = database
+        .get_block_children(&block_id, page_size, page)
+        .await
+        .map_err(|_| ApiError::NoResults)?;
+
+    let mut children = Vec::new();
+    while let Some(block_id) = block_children.try_next().await? {
+        children.push(block_id.to_hex());
+    }
+
+    Ok(BlockChildrenResponse {
+        block_id: block_id.to_hex(),
+        max_results: page_size,
+        count: children.len(),
+        children,
     })
 }
